+++ libs/framework-core/source/ftrack_framework_core/widget/dialog.py
@@ -83,6 +83,17 @@ class FrameworkDialog(BaseUI):
 
         self._tool_config = value
 
+        if self._tool_config:
+            arguments = {
+                "tool_config_reference": self._tool_config['reference'],
+                "plugin_config_reference": None,
+                "options": self.dialog_options,
+            }
+            self.logger.warning(f"arguments --> {arguments} ")
+            self.client_method_connection(
+                'set_config_options', arguments=arguments
+            )
+
         # Call _on_tool_config_changed_callback to let the UI know that a new
         # tool_config has been set.
         self._on_tool_config_changed_callback()
@@ -464,7 +475,7 @@ class FrameworkDialog(BaseUI):
         arguments = {
             "tool_config_reference": self.tool_config['reference'],
             "plugin_config_reference": plugin_reference,
-            "plugin_options": options,
+            "options": options,
         }
         self.client_method_connection(
             'set_config_options', arguments=arguments